namespace the Proposal query objects
