+++ app/policies/proposal_policy/scope.rb
@@ -12,7 +12,7 @@ class ProposalPolicy
       elsif @user.client_admin?
         self.for_client_admin
       else
-        @scope.where(Query::Proposals.which_involve(@user))
+        @scope.where(Query::Proposal::Clauses.which_involve(@user))
       end
     end
 
@@ -20,8 +20,8 @@ class ProposalPolicy
 
     def for_client_admin
       @scope.where(
-        Query::Proposals.for_client_slug(@user.client_slug).or(
-          Query::Proposals.which_involve(@user)
+        Query::Proposal::Clauses.for_client_slug(@user.client_slug).or(
+          Query::Proposal::Clauses.which_involve(@user)
         )
       )
     end
